// src/app/models/dependentdetailspersnolinfo.ts
export class ImmigrationdependentPersonalInfo {
    gender: string;
    dateOfBirth: string;
    ssn: string;
    alienRegistrationNumber: string;
    countryOfBirth: string;
    stateOfBirth: string;
    citizenship: string;
    facebook: string;
    linkedin: string;
    twitter: string;
    nationality: string;
}