// src/app/models/dependentdetailsprofile.ts
export class ImmigrationdependentProfile {
    fileNumber: string;
    status: string;
    creationDate: string;
    employeeId: string;
    email: string;
    firstName: string;
    middleName: string;
    lastName: string;
    phoneNumber: string;
    markForDeletion: string;
    deletedBy: string;
    deletedOn: string;
    relation:string
}